Feat: Created SortByButton component. Added SortByButton into /Dashboard
